+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -122,6 +122,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added a function (HDRenderPipeline.ResetRTHandleReferenceSize) to reset the reference size of RTHandle systems.
 - Added support for AxF measurements importing into texture resources tilings.
 - Added Layer parameter on Area Light to modify Layer of generated Emissive Mesh
+- Added support for camera relative ray tracing (and keeping non-camera relative ray tracing working)
 
 ### Fixed
 - Fix when rescale probe all direction below zero (1219246)